Move filter buttons out of map-control
